Move env translations into a new file
