+++ crates/nu-command/src/system/run_external.rs
@@ -7,7 +7,7 @@ use std::process::{Command as CommandSys, Stdio};
 use std::sync::atomic::Ordering;
 use std::sync::mpsc;
 
-use nu_engine::eval_block;
+use nu_engine::env_to_strings;
 use nu_protocol::engine::{EngineState, Stack};
 use nu_protocol::{ast::Call, engine::Command, ShellError, Signature, SyntaxShape, Value};
 use nu_protocol::{Category, Config, IntoInterruptiblePipelineData, PipelineData, Span, Spanned};
@@ -52,41 +52,10 @@ impl Command for External {
         let mut name: Spanned<String> = call.req(engine_state, stack, 0)?;
         let args: Vec<String> = call.rest(engine_state, stack, 1)?;
         let last_expression = call.has_flag("last_expression");
-        let env_vars = stack.get_env_vars();
 
         // Translate environment variables from Values to Strings
-        let mut env_vars_str = HashMap::new();
-        let config = stack.get_config()?;
-        for (env_name, val) in env_vars {
-            if let Some(conv) = config.env_conversions.get(&env_name) {
-                let block = engine_state.get_block(conv.to_string.0);
-
-                if let Some(var) = block.signature.get_positional(0) {
-                    let mut stack = stack.collect_captures(&block.captures);
-                    if let Some(var_id) = &var.var_id {
-                        stack.add_var(*var_id, val);
-                    }
-
-                    let val_str = eval_block(
-                        engine_state,
-                        &mut stack,
-                        block,
-                        PipelineData::new(Span::unknown()),
-                    )?
-                    .into_value(Span::unknown())
-                    .as_string()?;
-
-                    env_vars_str.insert(env_name, val_str);
-                } else {
-                    return Err(ShellError::MissingParameter(
-                        "block input".into(),
-                        conv.to_string.1,
-                    ));
-                }
-            }
-        }
-
         let config = stack.get_config().unwrap_or_default();
+        let env_vars_str = env_to_strings(engine_state, stack, &config)?;
 
         // Check if this is a single call to a directory, if so auto-cd
         let path = nu_path::expand_path(&name.item);
