+++ crates/nu-engine/src/lib.rs
@@ -1,7 +1,9 @@
 mod call_ext;
 mod documentation;
+mod env;
 mod eval;
 
 pub use call_ext::CallExt;
 pub use documentation::{generate_docs, get_brief_help, get_documentation, get_full_help};
+pub use env::{env_to_strings, env_to_values};
 pub use eval::{eval_block, eval_expression, eval_operator};
